Enum hdf5_derive::Error

pub enum Error {
    MissingDataset(MissingDataset),
    SerializeArray(SerializeArray),
    WriteArray(WriteArray),
    CreateDataset(CreateDataset),
    FetchDataset(FetchDataset),
    MissingAttribute(MissingAttribute),
    SerializeAttribute(SerializeAttribute),
    FetchAttribute(FetchAttribute),
    CreateAttribute(CreateAttribute),
    WriteAttribute(WriteAttribute),
    MissingGroup(MissingGroup),
    CreateGroup(CreateGroup),
}

General error type that provides helpful information on what went wrong

Variants

MissingDataset(MissingDataset)

Error when attempting to read a dataset that does not exist in an hdf5 file

SerializeArray(SerializeArray)

Error when attempting to read an array to a given struct type

This error can occur if the dimension of the h5 array does not match the dimension of the type provided

WriteArray(WriteArray)

Failed to serialize an array to a hdf5 dataset

CreateDataset(CreateDataset)

Could not create a dataset in a hdf5 file when writing

FetchDataset(FetchDataset)

Could fetch an existing dataset when writing file

MissingAttribute(MissingAttribute)

Attribute was missing from hdf5 file

SerializeAttribute(SerializeAttribute)

Attribute was missing from hdf5 file

FetchAttribute(FetchAttribute)

Attribute was missing from hdf5 file

CreateAttribute(CreateAttribute)

Could not create a attribute in a hdf5 file when writing

WriteAttribute(WriteAttribute)

Could not create a attribute in a hdf5 file when writing

MissingGroup(MissingGroup)

Could not open an existing group in a hdf5 file when writing

CreateGroup(CreateGroup)

Could not create a group in a hdf5 file when writing