Module hdf5_derive::error

Error types for handling HDF5 data with additional context information

Structs

CreateAttribute

Could not create a dataset in a hdf5 file when writing

CreateDataset

Could not create a dataset in a hdf5 file when writing

CreateGroup

Could not create a group in a hdf5 file when writing

DimensionMismatch

crate::lazy_array::LazyArray underlying [hdf5::Dataset] was not the correct dimension

FetchAttribute

Could fetch an existing dataset when writing file

FetchDataset

Could fetch an existing dataset when writing file

MissingAttribute

Could fetch an existing dataset when writing file

MissingDataset

Error when attempting to read a dataset that does not exist in an hdf5 file

MissingDatatype

Failed to fetch the datatype of a given dataset

MissingGroup

Could not open an existing group in a hdf5 file when writing

ReadSlice

Failed to read a slice of data from an HDF5 dataset

SerializeArray

Error when attempting to read an array to a given struct type.

SerializeAttribute

Failed to serialize the dataset to the correct type after it has been written

WriteArray

Failed to serialize an array to a hdf5 dataset

WriteAttribute

Failed to serialize an attribute to a hdf5 dataset

WriteSlice

Failed to write a slice of data to an HDF5 dataset

WrongDatatype

Datatype for crate::LazyArray was incorrect