Struct hdf5_derive::File

pub struct File(_);

HDF5 file object.

Implementations

impl File

pub fn open<P>(filename: P) -> Result<File, Error>where
P: AsRef<Path>,

Opens a file as read-only, file must exist.

pub fn open_rw<P>(filename: P) -> Result<File, Error>where
P: AsRef<Path>,

Opens a file as read/write, file must exist.

pub fn create<P>(filename: P) -> Result<File, Error>where
P: AsRef<Path>,

Creates a file, truncates if exists.

pub fn create_excl<P>(filename: P) -> Result<File, Error>where
P: AsRef<Path>,

Creates a file, fails if exists.

pub fn append<P>(filename: P) -> Result<File, Error>where
P: AsRef<Path>,

Opens a file as read/write if exists, creates otherwise.

pub fn open_as<P>(filename: P, mode: OpenMode) -> Result<File, Error>where
P: AsRef<Path>,

Opens a file in a given mode.

pub fn with_options() -> FileBuilder

Opens a file with custom file-access and file-creation options.

pub fn size(&self) -> u64

Returns the file size in bytes (or 0 if the file handle is invalid).

pub fn free_space(&self) -> u64

Returns the free space in the file in bytes (or 0 if the file handle is invalid).

pub fn is_read_only(&self) -> bool

Returns true if the file was opened in a read-only mode.

pub fn userblock(&self) -> u64

Returns the userblock size in bytes (or 0 if the file handle is invalid).

pub fn flush(&self) -> Result<(), Error>

Flushes the file to the storage medium.

pub fn close(self) -> Result<(), Error>

Closes the file and invalidates all open handles for contained objects.

pub fn access_plist(&self) -> Result<FileAccess, Error>

Returns a copy of the file access property list.

pub fn fapl(&self) -> Result<FileAccess, Error>

A short alias for access_plist().

pub fn create_plist(&self) -> Result<FileCreate, Error>

Returns a copy of the file creation property list.

pub fn fcpl(&self) -> Result<FileCreate, Error>

A short alias for create_plist().