Struct hdf5_derive::LazyArray

pub struct LazyArray<T, DIM>where
    DIM: Dimension,
    T: H5Type,{ /* private fields */ }

An object that can be treated similar to an ndarray::Array, but backed by an on-disk HDF5 dataset instead of an in-memory array.

LazyArray objects are type-checked and dimension-checked when constructed with LazyArray::new. LazyArrays are a nice abstraction over datasets that ensure dataset errors happen at the time of creation, not at the time of processessing. You should reach for a LazyArray when your dataset is otherwise too large to hold in memory. Otherwise, the convenience of using ndarray backed arrays is generally nicer than HDF5 backed arrays.

Macro Attributes

LazyArray ignores transpose and mutate_on_write attributes. crate::ContainerWrite’s write routines have no effect on a LazyArray, all data is written immediately to the underlying datasets.

transpose may have some effect in the future if required. For now, the following is ignored:

use hdf5_derive::ContainerRead;
use hdf5_derive::LazyArray5;

#[derive(ContainerRead)]
struct Container {
    #[hdf5(transpose="both")]
    large_array: LazyArray5<u64>
}

Types

T is the datatype of the underlying array (f32, f64, usize, etc), and DIM is the dimension of the data. Consider using a type alias such
as LazyArray3, LazyArray4, etc to skip specifying the dimension.

Implementations

impl<T, DIM> LazyArray<T, DIM>where DIM: Dimension, T: H5Type,

pub fn new(dataset: Dataset) -> Result<Self, Error>

generate a dimension-checked and type-checked LazyArray for a given dataset. Errors if the dimension of the dataset is different from the type, or if the type of the HDF5 data is different from the type.

Exmaple

let N = 500;
let path = "./some_h5_file.h5";
let file = hdf5_derive::File::create(path).unwrap();

// create a massive dataset, we would not want to hold this in memory
let dataset = file
    .new_dataset::<f64>()
    .shape((N, N, N, N))
    .create("some_dataset_name")
    .unwrap();

let lazy_array = hdf5_derive::LazyArray4::<f64>::new(dataset).unwrap();

std::fs::remove_file(path);

pub fn slice<I>(&self, info: I) -> Result<Array<T, I::OutDim>, Error>where I: SliceArg<DIM> + TryInto<Selection>, Error: From<I::Error>,

read a slice of data from the HDF5 dataset.

info is normally constructed with the ndarray::s macro, and slices into the dataset as it would slice into a ndarray::ArrayBase object. Usually this method does not fail as the dimension and type have been checked previously

Exmaple

use hdf5_derive::{LazyArray5, ContainerRead, ContainerWrite};
let path = "./another_file.h5";
let file = hdf5_derive::File::create(path).unwrap();

#[derive(ContainerWrite, ContainerRead)]
struct WriteHelper {
    arr: ndarray::Array5<f32>
}

let mut arr = ndarray::Array5::zeros((3,3,3,3,3));
arr[[1,2,1,1,2]] = 10.;

let helper= WriteHelper { arr: arr.clone() };
helper.write_hdf5(&file);
file.close();

#[derive(ContainerWrite, ContainerRead)]
struct LazyReader {
    arr: LazyArray5<f32>
}

let file = hdf5_derive::File::open(path).unwrap();
let lazy_reader = LazyReader::read_hdf5(&file).unwrap();

// now we can slice the data directly from the h5 file without loading the 
// entire file into memory!
let slice = lazy_reader.arr.slice(ndarray::s![1,2,1,1,2]).unwrap();

assert_eq!(slice.into_scalar(), 10.0);

std::fs::remove_file(path);

pub fn write_slice<'a, ARR, I>(&self, array: ARR, info: I) -> Result<(), Error>where ArrayView<'a, T, I::OutDim>: From<ARR>, I: SliceArg<DIM> + TryInto<Selection>, Error: From<I::Error>,

write an array slice to a the underlying HDF5 dataset

info is normally constructed with the ndarray::s macro, and slices into the dataset as it would slice into a ndarray::ArrayBase object. Usually this method does not fail as the dimension and type have been checked previously

Exmaple

use hdf5_derive::{LazyArray5, ContainerRead, ContainerWrite};
let path = "./another_file.h5";
let file = hdf5_derive::File::create(path).unwrap();

#[derive(ContainerWrite, ContainerRead)]
struct WriteHelper {
    arr: ndarray::Array5<f32>
}

let mut arr = ndarray::Array5::zeros((3,3,3,3,3));
arr[[1,2,1,1,2]] = 10.;

let helper= WriteHelper { arr: arr.clone() };
helper.write_hdf5(&file);
file.close();

#[derive(ContainerWrite, ContainerRead)]
struct LazyReader {
    arr: LazyArray5<f32>
}

let file = hdf5_derive::File::open_rw(path).unwrap();
let lazy_reader = LazyReader::read_hdf5(&file).unwrap();

// the data was the same as how we wrote it
let slice = lazy_reader.arr.slice(ndarray::s![1,2,1,1,2]).unwrap();
assert_eq!(slice.into_scalar(), 10.0);

// overwrite the data with a new slice
let write_val = ndarray::arr0(3.0f32);
let slice = lazy_reader.arr.write_slice(write_val.view(), ndarray::s![1,2,1,1,2]).unwrap();

// go back and read the data, it should be different now
let slice = lazy_reader.arr.slice(ndarray::s![1,2,1,1,2]).unwrap();
assert_eq!(slice.into_scalar(), 3.0);

std::fs::remove_file(path);

pub fn dataset(&self) -> &Dataset

provides access to the underlying dataset