Trait hdf5_derive::ContainerIo

pub trait ContainerIo {
    fn write_hdf5(&self, container: &File) -> Result<(), Error>;
    fn read_hdf5(container: &Group) -> Result<Self, Error>
    where
        Self: Sized;
}

Provides methods for reading and writing to an [hdf5] file. Derived with HDF5 macro.

Required Methods

fn write_hdf5(&self, container: &File) -> Result<(), Error>

write the contents of a struct to an HDF5 file

use hdf5_derive::ContainerIo;
use hdf5_derive::HDF5;
use ndarray::Array2;

#[derive(HDF5)]
struct Data {
    some_field: Array2<u32>
}

let path = "./test_file_write.h5";
let file = hdf5_derive::File::create(path).unwrap();

// write some data to the file
let arr = Array2::zeros((5,5));
let data = Data { some_field: arr.clone() };
data.write_hdf5(&file).unwrap();

// manually read the data using `hdf5` primitives
let dset = file.dataset("some_field").unwrap();
let read_array : Array2<u32> = dset.read().unwrap();

// check that they are the same
assert_eq!(read_array, arr);

// remove this file for practical purposes
std::fs::remove_file(path).unwrap();

fn read_hdf5(container: &Group) -> Result<Self, Error> where
Self: Sized,

read the contents of an HDF5 file to Self

use hdf5_derive::ContainerIo;
use hdf5_derive::HDF5;
use ndarray::Array2;

#[derive(HDF5)]
struct Data {
    #[hdf5(rename(read = "some_field_renamed"))]
    some_field: Array2<u32>
}

let path = "./test_file_read.h5";
let file = hdf5_derive::File::create(path).unwrap();

// write some data to the file
let arr = Array2::zeros((5,5));
let dset = file.new_dataset::<u32>()
    .shape((5,5))
    .create("some_field_renamed")
    .unwrap();
dset.write(&arr).unwrap();

// now, read the data from the written dataset
let read_data = Data::read_hdf5(&file).unwrap();

// check that they are the same
assert_eq!(read_data.some_field, arr);

// remove this file for practical purposes
std::fs::remove_file(path).unwrap();