Trait hdf5_derive::ReadGroup

pub trait ReadGroup {
    // Required method
    fn read_group(
        group: &Group,
        array_name: &str,
        transpose: bool
    ) -> Result<Self, Error>
       where Self: Sized;
}

Defines how a given piece of data should be parsed. You likely do not want to use this trait; instead use the methods from ContainerRead

Required Methods

fn read_group( group: &Group, array_name: &str, transpose: bool ) -> Result<Self, Error>where Self: Sized,

Given an hdf5 hdf5::Group and the name of the array we wish to read from it array_name, read the data from the array.