Trait hdf5_derive::WriteGroup

pub trait WriteGroup {
    // Required method
    fn write_group(
        &self,
        group: &Group,
        array_name: &str,
        transpose: bool,
        mutate_on_write: bool
    ) -> Result<(), Error>
       where Self: Sized;
}

Defines how a given piece of data should be written. You likely do not want to use this trait; instead use the methods from ContainerWrite

Required Methods

fn write_group( &self, group: &Group, array_name: &str, transpose: bool, mutate_on_write: bool ) -> Result<(), Error>where Self: Sized,

Given an hdf5 hdf5::Group and the name of the array we wish to read from it array_name, write the data to the correct dataset in the group