Module hdf5_derive::error

Structs

CreateAttribute

Could not create a dataset in a hdf5 file when writing

CreateDataset

Could not create a dataset in a hdf5 file when writing

FetchAttribute

Could fetch an existing dataset when writing file

FetchDataset

Could fetch an existing dataset when writing file

MissingAttribute

Could fetch an existing dataset when writing file

MissingDataset

Error when attempting to read a dataset that does not exist in an hdf5 file

SerializeArray

Error when attempting to read an array to a given struct type.

SerializeAttribute

Failed to serialize the dataset to the correct type after it has been written

WriteArray

Failed to serialize an array to a hdf5 dataset

WriteAttribute

Failed to serialize an attribute to a hdf5 dataset