Trait hdf5_derive::WriteGroup

pub trait WriteGroup {
    fn write_group(
        &self,
        group: &Group,
        array_name: &str,
        transpose: bool,
        mutate_on_write: bool
    ) -> Result<(), Error>
    where
        Self: Sized;
}

Defines how a given piece of data should be written. You likely do not want to use this trait; instead use the methods from ContainerWrite

Required Methods

fn write_group(
&self,
group: &Group,
array_name: &str,
transpose: bool,
mutate_on_write: bool
) -> Result<(), Error>where
Self: Sized,